Revert "Feat/7 permitir que usuarios acessem pagina de detalhes de cada pais"
