+++ angular.json
@@ -66,8 +66,7 @@
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "browserTarget": "countries-angular:build",
-            "proxyConfig": "proxy.conf.js"
+            "browserTarget": "countries-angular:build"
           },
           "configurations": {
             "production": {
