+++ package.json
@@ -2,9 +2,9 @@
   "name": "countries-angular",
   "version": "0.0.0",
   "scripts": {
-    "ng": "npx ng",
-    "start": "npx ng serve",
-    "debug": "npx ng serve --source-map",
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
@@ -19,16 +19,13 @@
     "@angular/platform-browser": "~9.1.13",
     "@angular/platform-browser-dynamic": "~9.1.13",
     "@angular/router": "~9.1.13",
-    "@fortawesome/angular-fontawesome": "^0.6.1",
-    "@fortawesome/fontawesome-svg-core": "^1.2.28",
-    "@fortawesome/free-solid-svg-icons": "^5.13.0",
     "rxjs": "~6.5.4",
     "tslib": "^1.10.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.901.13",
-    "@angular/cli": "^9.1.13",
+    "@angular/cli": "~9.1.13",
     "@angular/compiler-cli": "~9.1.13",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
