--- proxy.conf.json
@@ -1,11 +1,0 @@
-{
-  "/api": {
-    "target": "https://restcountries.com",
-    "secure": true,
-    "changeOrigin": true,
-    "pathRewrite": {
-      "^/api": "/v3.1"
-    },
-    "logLevel": "debug"
-  }
-}
