+++ src/app/app-routing.module.ts
@@ -1,14 +1,8 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import {CountryDetailsComponent} from './components/country-details/country-details.component';
-import {CountriesComponent} from './components/countries/countries.component';
 
 
-const routes: Routes = [
-  { path: '', redirectTo: '/countries', pathMatch: 'full' },
-  { path: 'countries', component: CountriesComponent },
-  { path: 'country/:name', component: CountryDetailsComponent }
-];
+const routes: Routes = [];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
