+++ src/app/app.component.html
@@ -1,2 +1,7 @@
 <app-header></app-header>
-<router-outlet></router-outlet>
+
+<app-filter (filterChanged)="onFilterChanged($event)"></app-filter>
+
+<section class="cards" *ngFor="let country of filteredCountries | slice:0:8">
+  <app-country-card [country]="country"></app-country-card>
+</section>
