+++ src/app/app.component.scss
@@ -3,3 +3,10 @@
   display: flex;
   flex-direction: column;
 }
+
+.cards {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 2.2rem;
+}
