--- src/app/components/countries/countries.component.scss
@@ -1,6 +1,0 @@
-.cards {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 2.2rem;
-}
