--- src/app/shared/constants/country-trigram-map.ts
@@ -1,196 +1,0 @@
-export const countryTrigramMap: Record<string, string> = {
-  // Europa
-  DEU: 'Germany',
-  FRA: 'France',
-  GBR: 'United Kingdom',
-  ESP: 'Spain',
-  ITA: 'Italy',
-  NLD: 'Netherlands',
-  BEL: 'Belgium',
-  CHE: 'Switzerland',
-  AUT: 'Austria',
-  PRT: 'Portugal',
-  GRC: 'Greece',
-  SWE: 'Sweden',
-  NOR: 'Norway',
-  DNK: 'Denmark',
-  FIN: 'Finland',
-  IRL: 'Ireland',
-  POL: 'Poland',
-  CZE: 'Czech Republic',
-  HUN: 'Hungary',
-  ROU: 'Romania',
-  BGR: 'Bulgaria',
-  HRV: 'Croatia',
-  SVK: 'Slovakia',
-  SVN: 'Slovenia',
-  LUX: 'Luxembourg',
-  LVA: 'Latvia',
-  EST: 'Estonia',
-  LTU: 'Lithuania',
-  BLR: 'Belarus',
-  UKR: 'Ukraine',
-  MDA: 'Moldova',
-  RUS: 'Russia',
-  MNE: 'Montenegro',
-  SRB: 'Serbia',
-  BIH: 'Bosnia and Herzegovina',
-  ALB: 'Albania',
-  MKD: 'North Macedonia',
-  AND: 'Andorra',
-  LIE: 'Liechtenstein',
-  MCO: 'Monaco',
-  SMR: 'San Marino',
-  VAT: 'Vatican City',
-
-  // América do Norte
-  USA: 'United States',
-  CAN: 'Canada',
-  MEX: 'Mexico',
-
-  // América Central e Caribe
-  BLZ: 'Belize',
-  CRI: 'Costa Rica',
-  SLV: 'El Salvador',
-  GTM: 'Guatemala',
-  HND: 'Honduras',
-  NIC: 'Nicaragua',
-  PAN: 'Panama',
-  CUB: 'Cuba',
-  DOM: 'Dominican Republic',
-  HTI: 'Haiti',
-  JAM: 'Jamaica',
-  BHS: 'Bahamas',
-  BRB: 'Barbados',
-  DMA: 'Dominica',
-  GRD: 'Grenada',
-  TTO: 'Trinidad and Tobago',
-
-  // América do Sul
-  BRA: 'Brazil',
-  ARG: 'Argentina',
-  BOL: 'Bolivia',
-  CHL: 'Chile',
-  COL: 'Colombia',
-  ECU: 'Ecuador',
-  GUF: 'French Guiana',
-  GUY: 'Guyana',
-  PRY: 'Paraguay',
-  PER: 'Peru',
-  SUR: 'Suriname',
-  URY: 'Uruguay',
-  VEN: 'Venezuela',
-
-  // Ásia
-  CHN: 'China',
-  JPN: 'Japan',
-  KOR: 'South Korea',
-  PRK: 'North Korea',
-  IND: 'India',
-  IDN: 'Indonesia',
-  MYS: 'Malaysia',
-  PHL: 'Philippines',
-  SGP: 'Singapore',
-  THA: 'Thailand',
-  VNM: 'Vietnam',
-  PAK: 'Pakistan',
-  BGD: 'Bangladesh',
-  AFG: 'Afghanistan',
-  IRN: 'Iran',
-  IRQ: 'Iraq',
-  SAU: 'Saudi Arabia',
-  SYR: 'Syria',
-  ARE: 'United Arab Emirates',
-  ISR: 'Israel',
-  JOR: 'Jordan',
-  KWT: 'Kuwait',
-  LBN: 'Lebanon',
-  OMN: 'Oman',
-  QAT: 'Qatar',
-  YEM: 'Yemen',
-  TUR: 'Turkey',
-  KAZ: 'Kazakhstan',
-  UZB: 'Uzbekistan',
-  TKM: 'Turkmenistan',
-  KGZ: 'Kyrgyzstan',
-  TJK: 'Tajikistan',
-  ARM: 'Armenia',
-  AZE: 'Azerbaijan',
-  GEO: 'Georgia',
-  CYP: 'Cyprus',
-  MNG: 'Mongolia',
-  MMR: 'Myanmar',
-  NPL: 'Nepal',
-  LKA: 'Sri Lanka',
-  BTN: 'Bhutan',
-  LAO: 'Laos',
-  KHM: 'Cambodia',
-  TLS: 'Timor-Leste',
-  MRT: 'Mauritania',
-  BHR: 'Bahrain',
-  ESH: 'Western Sahara',
-
-  // Oceania
-  AUS: 'Australia',
-  NZL: 'New Zealand',
-  PNG: 'Papua New Guinea',
-  FJI: 'Fiji',
-  SLB: 'Solomon Islands',
-  VUT: 'Vanuatu',
-  WSM: 'Samoa',
-  TON: 'Tonga',
-
-  // África
-  ZAF: 'South Africa',
-  EGY: 'Egypt',
-  NGA: 'Nigeria',
-  KEN: 'Kenya',
-  MAR: 'Morocco',
-  GHA: 'Ghana',
-  TZA: 'Tanzania',
-  DZA: 'Algeria',
-  TUN: 'Tunisia',
-  LBY: 'Libya',
-  SDN: 'Sudan',
-  ETH: 'Ethiopia',
-  UGA: 'Uganda',
-  AGO: 'Angola',
-  MOZ: 'Mozambique',
-  ZWE: 'Zimbabwe',
-  ZMB: 'Zambia',
-  NAM: 'Namibia',
-  BWA: 'Botswana',
-  SEN: 'Senegal',
-  MLI: 'Mali',
-  NER: 'Niger',
-  TCD: 'Chad',
-  CMR: 'Cameroon',
-  CIV: 'Ivory Coast',
-  GIN: 'Guinea',
-  BFA: 'Burkina Faso',
-  BEN: 'Benin',
-  GAB: 'Gabon',
-  COG: 'Republic of the Congo',
-  COD: 'Democratic Republic of the Congo',
-  RWA: 'Rwanda',
-  BDI: 'Burundi',
-  SOM: 'Somalia',
-  ERI: 'Eritrea',
-  SSD: 'South Sudan',
-  CAF: 'Central African Republic',
-  GMB: 'Gambia',
-  GNB: 'Guinea-Bissau',
-  LBR: 'Liberia',
-  SLE: 'Sierra Leone',
-  TGO: 'Togo',
-  MWI: 'Malawi',
-  LSO: 'Lesotho',
-  SWZ: 'Eswatini',
-  MDG: 'Madagascar',
-  MUS: 'Mauritius',
-  COM: 'Comoros',
-  SYC: 'Seychelles',
-  CPV: 'Cape Verde',
-  STP: 'São Tomé and Príncipe',
-  DJI: 'Djibouti'
-};
