+++ src/environments/environment.prod.ts
@@ -1,4 +1,3 @@
 export const environment = {
-  production: true,
-  apiUrl: 'https://restcountries.com/v3.1/api'
+  production: true
 };
